Remove spurious MPI send/recvs in halo expansion as well.
This is to avoid communication with all other processors with many zero
messages for large processor counts. Unfortunately this was a little
less clear to work out as in the expansion of the halo we do not yet
know which are our neighbouring processes included in the new expanded
halo.