Changes needed to build fluidity with petsc v3.8:
* the main difference in petsc's fortran interface  from v3.8 is the introduction of PETSC_NULL_XXX,
where XXX=KSP,Vec,Mat,etc. instead of the generic PETSC_NULL_OBJECT. Compatibility for older
petsc versions is maintained by #define-ing PETSC_NULL_XXX PETSC_NULL_OBJECT. Unfortunately a number
of bugs/complications were discovered associated with this change that I needed to work around:
- MatCreateSchurComplement is missing a hand-rolled fortran wrapper with a CHKFORTRANNULLOBJECT, so that
a PETSC_NULL_MAT is not correctly translated to a C (Mat *) null
- PetscObjectReference is also missing a wrapper (passing the %v member works instead)
- there is not PETSC_NULL_MATNULLSPACE and MatGetNullSpace is not wrapped correctly, workaround with
IsNullMatNullSpace() function that tests the returned nullspace
- in petsc 3.8, the check that petsc object arguments are null arguments (and thus meant to be omited/ignored)
is now checked by checking whether the value of the associated fortran derived type with single attribute %v,
has the value -1. It is therefore possible that uninitialised petsc object variables in fortran are uninintenionally
initialised to be null. This has unintended consequences for some petsc routines, such as MatCreateVecs, with
optional output arguments. If the provided uninitialised argument happens to contain -1 as input value, it will
be interpreted as null and therefore no output result is returned for it. Thus such arguments now need to be
initialized with a fortran petsc object that is known not to be equal to PETSC_NULL_XXX
- routines with missing wrappers to recognize PETSC_NULL_INTEGER: MatGetOwnerShipRange and MatSeqAIJSetPreallocation
- no such thing as PETSC_NULL_RANDOM, so VecSetRandom *needs* a Random object as input
- the (MPI_Comm *) comms arugment of PCMGSetLevels does not have a clear null option in fortran, PETSC_NULL_KSP seems to work
as there is actually no explcit wrapper
* remove petsc33 testharness tests as petsc <v3.4 is no longer supported and other small cleanups
associated with older petsc versions.
* removing some erronous cruft from configure
*  although a very simple test of petsc fortran modules was run in our configure (which prior to v3.8
decides whether the fortran modules are "functional", i.e. not compiled with a different fortran compiler)
the longer test was unintentionally not using modules. From petsc v3.8, petsc fortran modules are required
and the --disable-petsc-fortran-modules fluidity configure option is no longer supported (icw petsc 3.8)
* removing previous workarounds to do with bugs in MatGetInfo and PETSC_NULL_OBJECT
* PCGAMGSetThreshold is changed (and PCGAMGSetThreshold is added), need to discuss possibly better default for Stokes. For
now attempted to keep the same.