Some reorganisation to deal with derived CoordinateMesh
* Instead of temporarily reinserting the unpopped adapted coordinate field just
before calling interpolation, have the interpolation manager itself decide
to pull out the correct version of the coordinate field which is now
stored under SphericalAdaptivitiyBaseGeometry. Only temporarily
replacing the (adapted) coordinate field does not work if we have coordinate
fields that are derived, as the derived coordinate fields will already
be popped out at this stage.
* No longer store the 2 versions of the metric in state, as there is no
longer a use for this.