Add new write_extruded_mesh_only option
This option will change the extruded mesh to the external from_file mesh
and drop all horizontal meshes. It will still perform a columnar
decomposition based on the horizontal mesh. Rerunning flredecomp on the
output flml of this first step however, it will decompose in all
dimensions. This is a useful feature in setups where the extruded mesh
is too big to fit in memory on a single core.

This commit also fixes an issue with extrusion + arbitrary (i.e. not
vertically_structured) adaptivity. The checkpoints written by this setup
were using the same structure as a checkpoint with vertically structured
adaptivity, or non-adaptive extruded runs: i.e. the extruded (now
adapted mesh) is checkpointed and added under
extrude/checkpoint_from_file in the options tree. The checkpoint flmls
were functional for a restart (as tested in extrude_adapt3d_parallel/)
but could not be flredecomped, as the decomposition of the horizontal
and extruded mesh would be completely out of sync. To fix this, "normal"
adaptivity with extrusion now also changes the extruded (and since
adapted) mesh as the from_file mesh and removes the horizontal meshes
from the options tree. During the run the horizontal meshes are kept in
state, as they are needed to complete a consistent .stat file. The .stat
file from a pick up run is necessarily different if the original
(horizontal) InputMesh was included.

TODO: tests