Make stat file more robust for disappearing meshes.
With extrude+"normal" adapt we remove the horizontal meshes from the
options tree. If these were included in the .stat file that would create
problems as inclusion was checked for on every write. Now we only check
for inclusion once at the start of the run. For now, the horizontal meshes
are kept in state so that stat can keep writing their (unchanged)
statistics. If we do remove the horizontal meshes from state as well,
the stat file code should now also be able to handle it (it will then just
write zeros in the associated columns).

In extrude_adapt3d_parallel include both horizontal and extruded mesh.
Previously this would have failed. The case where only the horizontal
mesh is included only worked by fluke: after removal of the horizontal
mesh options, the include_in_stat option was switched on for the
extruded mesh and the columns for the horizontal mesh were taken over by
the extruded mesh by accident.