Archive output from failed tests
Acting on another suggestion from @jrper to archive test results. His comment in the issue:

Regarding point 3) in my wishlist above, and as a point of negotiation, in 7ab620a I've added a commit which updates test harness to have a new option -a or --archive-failures which takes a filename and writes directories which fail to gzipped tarball under that title. It also modifies the root makefile adding a new variable TESTHARNESS_OPTIONS to expose this.

Usage would be

make TESTHARNESS_OPTIONS="-a short_test_failures.tar.gz" test
make TESTHARNESS_OPTIONS="-a medium_test_failures.tar.gz" medium test

(with an added note that we may want to exercise fine-grained control on how long artifacts are retained by Jenkins)