Removing spud before adding it back with subtree
For future reference, the adding back will be done with:

     git subtree add -P libspud/ git@github.com:FluidityProject/spud.git master --squash

Afterwards, further updates to upstream spud can be pulled in using:

    git subtree pull -P libspud/ git@github.com:FluidityProject/spud.git master --squash

Or you can add spud first as a remote:

    git remote add -f spud git@github.com:FluidityProject/spud.git

and then pull using:

    git subtree pull -P libspud/ spud master --squash

The --squash is required as you can't combine squashed and non-squashed
commits for a subtree. The reason we use --squash in the first commit is
to avoid pullin the entire history of spud (which is unnecessarily big).