Specify base_geometry separately as option under /geometry/spherical_earh/base_geometry
This means spherical adaptivity no longer assumes an extruded mesh setup
for the initial mesh, where it reads the base geometry from the
horizontal "InputMesh". This should fix checkpointing issues.

The base_geometry mesh specified  under
/geometry/spherical_earh/base_geometry will not be (re)decomposed and
thus always points to a serial input mesh, which will be read by process
0 and distributed among the cores. The radius of the base_geometry mesh
is ignored - instead a top and bottom version of the base_geometry mesh
will be constructed (based on the same mesh topology) with the radius
determined by the corresponding nodes in the top and bottom surface
meshes of the computational mesh.

Also fix spherical adaptivity for 2D: the newly added check for
inversion of elements now also works in 2D.