Improve comments of face_n_s and face_dn_s
Rather than reordering n_s and dn_s to match the facet ordering (which
is what it claimed to do), I think it's clearer to say that we assume
that the precomputed n_s and dn_s are already consistent with the facet
ordering (in terms of its gauss points and local coordinate
derivatives). It's actually the ordering of the element basis functions
(which in the precomputation was assumed to be such that basis function
1 is associated with the node opposite the facet, etc. that is changed
to match the ordering of the actual element node ordering in the mesh.
In other words it's the element that needs reordering, not the facet.