Update AC_PYTHON_DEVEL aclocal macro
Versions of Python 3 may be tagged according to their ABI version (PEP
3149), e.g. `3.7m', however the current AC_PYTHON_DEVEL macro is
unable to detect this, leading to a discrepancy between PYTHON_VERSION
and the libraries it attempts to link against. This updates the macro
to the latest aclocal-contrib version, while retaining the custom
behaviour for OSX in 139814b57 ("Minimal changeset to configure and
build successfully with python 3.x.", 2018-03-07).