Convert and add tests to for particle format
This converts the particle test suite to use the H5hutpy module to
read the particle output files. Tests previously using Lagrangian
detectors now use particles, and a few more tests have been added.

Support setting particle group output frequency

Similar to the /io/dump_period and /io/dump_period_in_timesteps
options, we can set the per-group particle output frequency. If
neither of these options are specified, it is assumed that particles
output for the group is desired at every timestep.

Accompanying this change, the output file format changes slightly:
instead of following the numeric timesteps, steps in the output file
are sequential, regardless of the dump period. To recover timestep
information, an integer "timestep" attribute is added to steps in the
output file.

Save particle IDs in dumps and checkpoints

The older IO format would save particles in ID-order, so their order
was preserved regardless of the decomposition. With H5hut, the
particles are written to file batched by processor, which means that
the order is no longer preserved. We save the (as of now)
globally-unique particle IDs to dumps and checkpoints so that
individual particles can be tracked.