Support setting particle group output frequency
Similar to the /io/dump_period and /io/dump_period_in_timesteps
options, we can set the per-group particle output frequency. If
neither of these options are specified, it is assumed that particles
output for the group is desired at every timestep.

Accompanying this change, the output file format changes slightly:
instead of following the numeric timesteps, steps in the output file
are sequential, regardless of the dump period. To recover timestep
information, an integer "timestep" attribute is added to steps in the
output file.