Convert gmsh transformation tools to use meshio
To support msh2 and msh4 formats transparently, the two Python tools
which are used in the test suite have been altered to use the meshio
library. This doesn't quite have full support for version 4.1, but
seems like a nicer solution than manually modifying the msh files
anyway. For the shelf_meshtransform_freesurfaceinitialcondition test
case, something goes a bit wonky in meshio, so it has been forced to
used a msh2 input.