Fix checkpointing.
Previously the code assumed that the input mesh was popped in already.
This is only true when starting from an input mesh that is equal to the
base geometry. When checkpointing, or when starting with an input mesh
that is finer than the base geometry, we need to pop it in to conform to
the base geometry.

Also change the base_geometry_before debugging vtu to use
vtk_write_fields instead of vtk_write_state, to ensure it actually uses
the right coordinate field (in case of superparametric for instance).