Perform (big) L2 nullspace projection after vector solve
With remove_null_space, a (small) l2 projection is performed at each
iteration projecting out the null space components, but in particular
for non-uniform meshes the l2-projection and L2-projection can differ
significantly leading to unnatural rotations/constant velocity
solutions. In the case of vector (velocity) solve this is now projected
out after the solve. Although the same would make sense for scalar
solves (with a constant nullspace), this is unfortunately a little
harder to achieve as petsc_solve() does not always get passed the
positions field (only if it goes via petsc_solve_state).

Also remove the silly, and unused dot_product for scalar/vector fields
to be replaced by a L2 dot_product for vector fields.