Move and reimplement L2 projection of vector nullspace
to outside the petsc_solve in the solvers module to the end of
momentum_equation. This is because in the former location the solution
(and the nullspace in the form of a MatNullSpace) are still in rotated
form. This means we cannot compute the integrals required for a L2
projection. To simplify matters (because I don't feel like untangling
the whole rotating back and forth mess that is going on in momentum
equation), the projection is only done once right at the end, i.e. only
after the final velocity correction - which I think is the only place we
care about it.