Handle optional Entities in GMSH4 files
The msh4 format made the $Entities section optional. We had been using
this section to get information about physical elements (e.g. for
boundary conditions). When meshio writes a file (such as when using
transform_mesh), it doesn't write the $Entities section, and instead
uses $ElementData to denote the physical element information.

This change introduces an alternate logic path, where if the $Entities
section is optional, we assume the msh file has been written by
meshio, and look instead for $ElementData. This is leaning pretty
heavily on the assumption that $ElementData is formatted only by
meshio.