Fix spher. adapt + galerkin projection
Galerkin projection should skip the coordinate field for interpolation.
In spherical adaptivity we use SphericalBaseGeometry as
our coordinate field which is stored under "Coordinate" in the
intermediate state that galerkin projection sees, but still has a field%name
different than "Coordinate". Now it checks for the name the vector field
is under rather than the field%name whether to skip it for
interpolation.