Remove last bit of UTF-8 from .bib
We should probably switch to biblatex+biber for this. Bibtex does not
handle utf8 correctly, just treats everything as a byte string, which
means it may cut things in the middle of a utf8 character. Or, in this
case, a warning about the fonts not supporting a character produced
broken utf8 in the fluidity_manual.log which breaks the `manual` test.