Remove erronous scaling of extruded partioning weights
This is in the partioning of the horizontal mesh. The weight for each
column is determined by the number of layers in that column. This weight
was then divided by the maximum number of layers, however the maximum
was not computed globally, so this scaling causes an imbalance between
partititions with many layers and those with few layers. Establishing
the global maximum (allmax) would probably cause a hang as this is inside a zoltan callback. Therefore just remove the scaling. This might give a different weighting of nodal vs. edge weights - but it's unclear what is the right answer.

Fixed #114.