Fix wrong Jacobian dimensions in transform_cvsurf_to_physical.
The second dimension of J should be the surface dimension (i.e. dim-1),
otherwise the matmul in line 2326 does not match. This is copying a
smaller array into a larger, so should be safe - and I can't actually
get my compiler to complain (even with debugging) - but this should
hopefully fix #206.