Flush particle file step after writing
If fluidity is interrupted before closing the particle file, in
parallel, the particle file tends to be corrupted in a way that can't
be read. If we make sure to flush the file after writing each step, we
reduce the likelihood of this happening. It is, however, hard to
guarantee that it strictly won't happen without proper transactional
writes in HDF5.