Prevent unallocated/unassociated actual arguments
In a few places, an array will be allocated or a pointer associated
depending on a logical flag. These subroutines then call another
subroutine, passing the array/pointer and the flag. However, the dummy
argument to this second subroutine doesn't have the allocatable or
pointer attributes. Depending on the flag, we may be passing an
unallocated array or unassociated pointer to a bare dummy argument,
which is not allowed. To be compliant, we should specify these
attributes, even for intent(in) dummy arguments, where the
allocated/associated status is not allowed to change (this is a
Fortran 2003-ism in the case of pointers).