Adding subroutines to zero the bubble dof in fields of different rank and doing this for rotational null spaces.
Rotational null spaces on bubble function spaces previously just evaluated a rotational component at the cell center for the bubble.
This isn't actually a representation of a rotation anymore.  The dof associated with the bubble should be zero in this case so this
commit does just that.

Currently the subroutines for zeroing the bubble just assume that the final dof in an element is the bubble.  Can't see any better
logic than this but if element numbering ever changes this would have to change too.