Speed up local_coords_interpolation()
used in a.o. lagrangian detector code. Speeding up for the case that the
coordinate mesh is linear. Instead of inverting a matrix to obtain the
local coordinates of a given physical location, uses the cached inverse
jacobian. Moves local_coords_interpolation to transform_elements as it
depends on its cache.