Reenable old single layer extrusion options.
This allows specifying extrusion options in two ways:
* single layer with one (WholeMesh) or more (named) regions directly
under extrude/. This is the same as the extrusion options in master and
therefore allows all current (single layer) flmls to remain unchanged
* one or more layers specified under extrude/ with regions under each
layer. This is the new layered option introduced in this branch

The schema allows both layers and regions directly under extrude/ but
this is disallowed through an options check. In diamond the user chooses
between regions and layers through a dropdown, with regions being the
default as I intend to maintain this as the simpler standard way to set
things up in the common single layer case.

This commit does not change back any of the single layer cases yet to
what they were on master, as I first want to test whether everything
still works when setup as a single (multi-)layer case on buildbot.