Adds capability for unittestharness to display the root directory of the unittest in the generated junit report.
This is only relevant for junit report generation.

The unittestharness attempts to get the root directory's name
e.g. error_metrics, adaptivity, using regex and trying to locate the
part of the test's path that is contained between `/tests/` and the
directory before that. 2 different approaches are used, either by
following the test's symlink or by deducing it from the absolute path.
If both fail a default name is used.

NOTE:
Both these approches require for the unittests to be placed in
a `tests/` subdirectory within each root directory, pretty standard stuff