Added the capability for unittestharness to: - run individual tests. - run a batch of tests from a file.
This commit deprecates the tool `runut.sh` and will be removed soon.

NOTE: Neither of the added options performs checks that the input test
name(s) are valid. Hence if invalid unit test names are provided the
unittestharness will simply ignore them and move to the next test.

NOTE 2: The unittestharness will search for the test(s) supplied by
the -t or --from-file options in the directory specified by -d which by
default is bin/tests.