BUG-FIX: Remove "import commands" from plotting scripts
This module is part of the stdlib in python2 only. Other than importing
it seemed to not actually be used. The plotting scripts
were called in the running of the tests but their output is not tested,
so the failure to import (in python3) only results in spurious errors
(python tracebacks) in the mediumtest output.