Fix extrusion with unspecified region ids
This is for the case where there are region ids specified in the options
and the input mesh also does not have region ids.

Bug introduced in bd623421d0168e481a76b9a5a5b335bc913f5d04:
although the new mesh%regions was initialised to 0, it then gets
clobbered by an uninitalised extruded_region_ids array that determines
the region id in each column.

This should actually be backported to master.