Bug-fix for extruded meshes
This is for the case where horizontal mesh appears before the extruded mesh in the options tree and
the horizontal mesh is removed from options. Because the removal
happened before the extruded mesh options where updated, the option path
that is indexed by number to the extruded mesh was no longer valid. Now
the horizontal mesh options are removed after the update of the extruded
mesh.