Moves the VTK dump after a solve.
This means that the solution is now displayed on the mesh on which it
occured and not on the interpolated mesh after a spatial adapt.

A `simulation_completed(current_time, timestep+1)` had to be used
for the cases where the option of `final_timestep` has been given to
prevent double output in the last timestep.

This still leaves the `adapt_at_first_timestep` since by the time the
first data dump occurs the mesh has already adapted, hence the 0th
and 1st timestep dumps look the same mesh-wise.