Selectively add -fallow-interface-mismatch
Stricter interface checking in gfortran10 is not going to work with
netcdf interface thus we need to disable that check. For now only in
ocean_forcing/ with gfortran >=10

Adds new AX_COMPARE_VERSION macro to compare versions more easily and
introduces new @FCFLAGS_ALLOW_INTERFACE_MISMATCH@  that get substituted
to -fallow-interface-mismatch with gfortran>=10 and is empty otherwise
so we can selectively apply this flag in Makefiles.