Adding initial CI support for GitHub Actions
This commit adds a workflow to build and run unit and short tests on
Ubuntu Bionic via GitHub Actions, along with a supporting Dockerfile
for the container in which the build and tests run.

Mediumtests are not included in this commit as they do not reliably
complete within the 6h time limit for GitHub builders and will either
require the test suite slimming down or self-hosted builders being
attached.

Note: testing is currently restricted to the branch this commit
initially went to (gcc10-fixes) and this will need to be updated
before merge, if this commit is included in a merge.