Adding initial CI support for GitHub Actions
This commit adds a workflow to build and run unit, shor, and medium
tests on Ubuntu LTS (bionic and focal) and latest release (groovy)
via GitHub Actions, along with a supporting Dockerfile for the
container in which the build and tests run.

The base image Dockerfiles are also overhauled to speed up CI run
times.

All three Ubuntu runs are condensed into one workflow as the JUnit
publishing stage publishes directly to the repo and Actions only
shows the published reports on one workflow.

Note: testing is currently restricted to the branch this commit
initially went to (gcc10-fixes) and this will need to be updated
before merge, if this commit is included in a merge.
Fixing paths to artifacts, both in upload and reading.