CI/Actions: adding focal support, dockerfile updates
This commit adds Actions support for testing on focal, and updates
the baseimage dockerfiles to bring them in line with the requirements
for Actions support, catching up with previous commits.

As with other Actions workflows, the 'builds on' is targetted to this
(gcc10-fixes) branch initially and should be reconfigured before merging.