TESTS: Reassign test lengths by current runtime
The medium test suite has grown to be extremely large, to the point
that it doesn't fit in hosted CI runs any more. A significant amount
of this appears to be down to a large number of short-running tests
which by their runtime (sub-300s) should be in short tests. This
commit reassigns them to short testing, which increases the short
testing total runtime but critically gets the medium testing total
runtime back comfortably under 6h (~4.5h).