Ensure write_metric_stages debugging output uses metric mesh
In cases where the coordinate mesh is not the same as the metric mesh
(higher order coordinates, where the metric is on the linear mesh) some
of the debugging output (edgelengths) was calculated on the wrong mesh.
Also many for many of the associated vtk outputs files the output mesh
was set to positions%mesh but should be metric%mesh.