CI/Actions: Separate out test reports
This commit changes behaviour back to failing a test job if the test
run and/or the tests themselves experience a failure. The benefit of
this is being able to much more precisely pinpoint where a failure
is in terms of the test suite it comes from, at the point of the
junit parsing.