CI/Actions: Add some longtesting
Adding in some longtesting, intended for tests which will
run on Github Actions runners (sub-6h). This is on the order
of the medium test suite duration so shouldn't slow down the
overall run.

Initial commit adds in longtests which are expected to pass;
other tests can be added over time as they are fixed and
deemed still useful.