Improved error messages from python
Often when an error occurs in the python embedded in fluidity, the
error+backtrace would not show up despite use calling PyErr_Print and
only print useless FLAbort message "Dying!". This
is caused by some internal buffering in python of stdio and stderr
(internal to python so independent of whether we have redirected the
actual stdio/stderr of the process at the system level). This commit
switches all such buffering off. As suggested here:
https://stackoverflow.com/questions/29352485/python-print-not-working-when-embedded-into-mpi-program/49076389#49076389
under option 1) Note that it should only affect stdio/err and not all
I/O from python (as it claims on stackexchange), and since we
shouldn't normally be printing much from python this should hopefully
not affect performance.

Also fixed a case where the error would result in a segfault because we
try to decref the pyobject returned from PyRun_String before checking
whether an error has occured, in which case PyRun_String returns null.

This came up in #315