Improve edge splitting in libadaptivity:
Previously an edge split was only performed if it lead to immediate
improvement with the new node half way the split edge. Now, when
evaluating potential edge splits if the split doesn't immediately lead
to an improvement we also consider moving the node to a more optimal
location (using the same node movement algorithm as for exisiting nodes)
and accept the edge split if it leads to improvement with the new node
on the optimised location. This should hopefully give libadaptivity more
room to fix very flat elements under a curved surface.