Fix detectors with moving mesh (free surface).
* Since we transform the physical domain back to its original shape
  during an adapt, but didn't update the physical position of detectors
they ended up outside of their element but more importantly can end up
outside the transformed physical (either local or global) domain. In
that case we aren't able to find back a containing element in the new
mesh for all detectors after an adapt. Now, the detectors are also moved
(transformed back) with the mesh before the adapt, and moved back
afterwards. This is done by keeping the local coordinates fixed and
reinterpolating to a position within the element. Once the element has
been transformed back this should gives us back the original pre-adapt
detector position.
* Deleted the second half of search_detectors as it has an incorrect and
  potentially dangerous call to picker_inquire with global=.true. This
combination only works if we provide picker_inquire with the same
positions to search on all processes, whereas search_for_detectors
naively assumed each process could just ask for its own missing detector
locations and picker_inquire would somehow figure it out. If in the
first (and now only) call to picker_inquire with global=.false. not all
detectors have been found, i.e. some detectors have left the local part
of the physical domain that we can see, this should be picked up in a
subsequent call to distribute_detectors(). In the case where we are
absolutely sure that this doesn't (shouldn't) happen (i.e. all detectors
can be found locally), we should check that search_detectors() indeed
has found all detectors. This happens in adapt, where the local physical
domain does not change during the adapt. We now check this in all runs
(not just with --enable-debugging) as it is cheap to do and we will always
crash in zoltan if we have lost detectors.