Fix "mg" preconditioner with petsc 3.14
This is dealing with bug https://gitlab.com/petsc/petsc/-/issues/868
Which is fixed for 3.15. Unfortunately no workaround exists in fortran,
because even if we do provide the optional argument it still gets
recognized for NULL and hits the bug. So instead write our own wrapper
in C.

This also drop support for petsc version <3.8 Note that Bionic official
packages are 3.7 - but we have always supplied our own 3.8 package for
it. Groovy and Hirsute are now on 3.14 Dropping support for <3.8
simplifies/removes a lot of the legacy workarounds. We no longer support
linking with petsc without fortran modules (we never did for versions
>=3.8).