Revert changed to external libraries
For h5hut, libjudy, libmba2d/3d, libspud and spatialindex we only
include the source for convenience. We may make some bespoke changes but
like to stay as close as possible to upstream.

libadaptivity and libvtkfortran no longer have an upstream
and we simply treat these subdirectories as the cannonical version, so
these are still included in the pre commit hook.